+++ crates/alien_provider/src/error.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone, thiserror::Error, uniffi::Error)]
+pub enum AlienError {
+    #[error("Request is invalid: {msg}")]
+    RequestError { msg: String },
+    #[error("Request error: {msg}")]
+    ResponseError { msg: String },
+}
