+++ crates/alien_provider/src/provider.rs
@@ -1,4 +1,5 @@
-pub use super::{AlienError, AlienTarget};
+pub use crate::{AlienError, AlienTarget};
+
 use async_trait::async_trait;
 use primitives::Chain;
 use std::fmt::Debug;
