+++ crates/alien_provider/src/target.rs
@@ -1,12 +1,27 @@
-use std::{collections::HashMap, fmt::Debug};
+use std::fmt::Debug;
 
 pub const X_CACHE_TTL: &str = "x-cache-ttl";
+
+#[derive(Debug, Clone, uniffi::Record)]
+pub struct AlienHeader {
+    pub key: String,
+    pub value: String,
+}
+
+impl AlienHeader {
+    pub fn new(key: &str, value: &str) -> Self {
+        Self {
+            key: key.into(),
+            value: value.into(),
+        }
+    }
+}
 
 #[derive(Debug, Clone, uniffi::Record)]
 pub struct AlienTarget {
     pub url: String,
     pub method: AlienHttpMethod,
-    pub headers: Option<HashMap<String, String>>,
+    pub headers: Option<Vec<AlienHeader>>,
     pub body: Option<Vec<u8>>,
 }
 
@@ -24,23 +39,29 @@ impl AlienTarget {
         Self {
             url: url.into(),
             method: AlienHttpMethod::Post,
-            headers: Some(HashMap::from([("Content-Type".into(), "application/json".into())])),
+            headers: Some(vec![AlienHeader {
+                key: "Content-Type".into(),
+                value: "application/json".into(),
+            }]),
             body: Some(serde_json::to_vec(&body).unwrap()),
         }
     }
 
     pub fn set_cache_ttl(mut self, ttl: u64) -> Self {
         if self.headers.is_none() {
-            self.headers = Some(HashMap::new());
+            self.headers = Some(vec![]);
         }
         if let Some(headers) = self.headers.as_mut() {
-            headers.insert(X_CACHE_TTL.into(), ttl.to_string());
+            headers.push(AlienHeader {
+                key: X_CACHE_TTL.into(),
+                value: ttl.to_string(),
+            });
         }
         self
     }
 }
 
-#[derive(Debug, Clone, PartialEq, uniffi::Enum)]
+#[derive(Debug, Clone, Copy, PartialEq, uniffi::Enum)]
 pub enum AlienHttpMethod {
     Get,
     Post,
@@ -64,9 +85,4 @@ impl From<AlienHttpMethod> for String {
         }
         .into()
     }
-}
-
-#[uniffi::export]
-fn alien_method_to_string(method: AlienHttpMethod) -> String {
-    method.into()
 }
