+++ crates/gem_evm/Cargo.toml
@@ -18,6 +18,7 @@ rpc = [
 primitives = { path = "../primitives" }
 gem_hash = { path = "../gem_hash" }
 serde_serializers = { path = "../serde_serializers" }
+gem_jsonrpc = { path = "../gem_jsonrpc" }
 
 hex = { workspace = true }
 itoa = { workspace = true }
