+++ crates/gem_solana/src/jsonrpc.rs
@@ -1,6 +1,8 @@
 use std::fmt::Display;
 
-use crate::model::Filter;
+use crate::model::{Configuration, Filter};
+use primitives::jsonrpc_types::{JsonRpcRequest, JsonRpcRequestConvert};
+use serde_json::Value;
 
 pub const ENCODING_BASE64: &str = "base64";
 pub const ENCODING_BASE58: &str = "base58";
@@ -22,5 +24,28 @@ impl Display for SolanaRpc {
             SolanaRpc::GetEpochInfo => write!(f, "getEpochInfo"),
             SolanaRpc::GetLatestBlockhash => write!(f, "getLatestBlockhash"),
         }
+    }
+}
+
+impl JsonRpcRequestConvert for SolanaRpc {
+    fn to_req(&self, id: u64) -> JsonRpcRequest {
+        let val = self;
+        let method = val.to_string();
+        let default_config = Configuration::default();
+
+        let params: Vec<Value> = match val {
+            SolanaRpc::GetProgramAccounts(program, filters) => vec![
+                Value::String(program.into()),
+                serde_json::to_value(Configuration::new(filters.to_vec())).unwrap(),
+            ],
+            SolanaRpc::GetAccountInfo(program) => vec![Value::String(program.into()), serde_json::to_value(default_config).unwrap()],
+            SolanaRpc::GetMultipleAccounts(accounts) => vec![
+                Value::Array(accounts.iter().map(|x| serde_json::to_value(x).unwrap()).collect()),
+                serde_json::to_value(default_config).unwrap(),
+            ],
+            SolanaRpc::GetEpochInfo | SolanaRpc::GetLatestBlockhash => vec![],
+        };
+
+        JsonRpcRequest::new(id, &method, params)
     }
 }
