+++ crates/primitives/src/jsonrpc_types.rs
@@ -1,0 +1,68 @@
+use serde::{Deserialize, Serialize};
+use std::fmt::{Debug, Display};
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct JsonRpcRequest {
+    pub jsonrpc: &'static str,
+    pub id: u64,
+    pub method: String,
+    pub params: Vec<serde_json::Value>,
+}
+
+pub trait JsonRpcRequestConvert {
+    fn to_req(&self, id: u64) -> JsonRpcRequest;
+}
+
+impl JsonRpcRequest {
+    pub fn new(id: u64, method: &str, params: Vec<serde_json::Value>) -> Self {
+        Self {
+            jsonrpc: "2.0",
+            id,
+            method: method.into(),
+            params,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct JsonRpcError {
+    pub code: i32,
+    pub message: String,
+}
+
+impl Display for JsonRpcError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{} ({})", self.message, self.code)
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct JsonRpcResponse<T> {
+    pub id: u64,
+    pub result: T,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct JsonRpcErrorResponse {
+    pub id: u64,
+    pub error: JsonRpcError,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+#[serde(untagged)]
+pub enum JsonRpcResult<T> {
+    Value(JsonRpcResponse<T>),
+    Error(JsonRpcErrorResponse),
+}
+
+impl<T> JsonRpcResult<T>
+where
+    T: Clone,
+{
+    pub fn take(&self) -> Result<T, JsonRpcError> {
+        match self {
+            JsonRpcResult::Value(value) => Ok(value.result.clone()),
+            JsonRpcResult::Error(error) => Err(error.error.clone()),
+        }
+    }
+}
