+++ crates/primitives/src/lib.rs
@@ -124,3 +124,5 @@ pub mod transaction_id;
 pub use self::transaction_id::TransactionId;
 pub mod asset_address;
 pub use self::asset_address::AssetAddress;
+
+pub mod jsonrpc_types;
