+++ gemstone/Cargo.toml
@@ -23,6 +23,7 @@ gem_sui = { path = "../crates/gem_sui" }
 gem_aptos = { path = "../crates/gem_aptos" }
 gem_hash = { path = "../crates/gem_hash" }
 serde_serializers = { path = "../crates/serde_serializers" }
+alien_provider = { path = "../crates/alien_provider" }
 
 bcs.workspace = true
 sui-types.workspace = true
