+++ gemstone/src/sui/rpc/mod.rs
@@ -1,5 +1,4 @@
 pub mod client;
-pub mod jsonrpc;
 pub mod models;
 
 pub use client::SuiClient;
